--- server/modules/apartment/utils/parse-rooms.ts
@@ -1,3 +1,0 @@
-export default function parseRooms(roomsStr: string): number[] {
-  return roomsStr.split(",").map(r => Number(r.trim()));
-}
